requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
chardet==3.0.4
charset-normalizer==2.0.6
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
distlib==0.3.3
Django==3.2.7
django-cors-headers==3.8.0
django-multiselectfield==0.1.12
djangorestframework==3.12.4
filelock==3.3.2
hiredis==2.0.0
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
msgpack==1.0.2
multidict==5.1.0
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.22.0
service-identity==21.1.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
supervisor==4.1.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.25.11
virtualenv==20.0.17
websockets==10.0
yarl==1.6.3
zope.interface==5.4.0
